requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Werkzeug==2.3.7
pytz>=2023.3
reportlab==4.0.4
requests==2.31.0
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<2.1.0
openpyxl==3.1.2